Increase example-seed-data.json list to 4096 words
This helps the user in two ways:

    1. It provides exactly 12-bits of entropy per word (2^12 = 4096) if they
       want to use this as their word list.
    2. It helps encourage users to deploy large word lists for increasing
       security.

Using this word list, this is the amount of entropy you get in the passphrase
as a result:

    * 4 words: 48-bits (weak)
    * 5 words: 60-bits (medium)
    * 6 words: 72-bits (strong)

This word list is a sorted list of the most common 4,096 English words that are
either 4, 5, or 6 characters in length. This will help minimize the number of
characters in the passphrase, without compromising entropy.

Example 4-word passphrases:

    * trade-what-senior-used (22 characters)
    * shack-drugs-curse-tables (23 characters)
    * write-faint-winner-alvin (23 characters)

Example 5-word passphrases:

    * memory-ichi-iove-olive-arms (27 characters)
    * fate-angus-enemy-parent-writes (30 characters)
    * nigel-serial-site-bomb-bait (27 characters)

Example 6-word passphrases:

    * rays-buys-brings-marc-tribe-trust (33 characters)
    * foot-snakes-bags-this-eighty-sylvia (35 characters)
    * marco-arms-dusty-route-stayed-vest (34 characters)