Increase default word count in passphrase 6
With example-seed-data.json now 4096 words in length, if that is used for the
word list, then 12-bits of entropy are provided for each word. This increases
the entropy of the system to 72-bits, which is considered secure.

To motivate this, considered the following scenarios:

1. The Bitcoin mining network, using specially designed ASICs is brute-forcing
SHA-256 hashes. The current pace is [7,430,760.76
TH/s](https://blockchain.info/stats). That's about 7.5x10^18 SHA-256 hashes, or
every combination in 62-bits calculated every second. It takes the Bitcoin
mining network just under an hour to calculate every combination of 74-bits.

2. The distributed.net distributed computing project is working on cracking a
[72-bit RC5 key](https://stats.distributed.net/projects.php?project_id=8). The
sustained rate is around 800 million keys per second, or calculating all of
29-bits each second. At that pace, they can only reach 54-bits per year, which
means it would take them around 130 thousand years at that pace to reach 50% of
the key space, thus having the probability of at least 50% of finding the right
key.

3. Recently, Troy Hunt released 320 million password hashes to the Internet. A
group of password crackers armed with MDXfind and Hashcat set to work. At peak,
they were [calculating 180 billion hashes per
second](https://cynosureprime.blogspot.com/2017/08/320-million-hashes-exposed.html).
That's 27-bits per second, 52-bits per year, and around 500 thousand years to
reach 71% (half the size of 72-bits).

I think it's fair to say that the Bitcoin mining power is at the very cusp of
what several well-funded state adversaries could afford in a targeted attack.
The rate of distributed.net looking for the 72-bit RC5 key or a cluster of 25
Nvidia GPUs is much more realistic for password cracking. As such, 48-bits is
well within reach of full compromise, 60-bits offers a small margin of
security, while 72-bits is clearly out of range for the time being.

As such, it's considered best practice for password generators to be generating
passwords with at least 70-bits of entropy.